Reducers have now Changes regarding a package refactor (now actions file are in the root store package).
