+++ store/src/main/kotlin/com/cesarvaliente/kunidirectional/store/Actions.kt
@@ -17,10 +17,7 @@
  * limitations under the License.
  */
 
-package com.cesarvaliente.kunidirectional.store.action
-
-import com.cesarvaliente.kunidirectional.store.Color
-import com.cesarvaliente.kunidirectional.store.Item
+package com.cesarvaliente.kunidirectional.store
 
 sealed class Action
